02.09.2020 11:31

+++ WpfApp5/MeasurementUnitDefination.xaml.cs
@@ -40,60 +40,66 @@ namespace WpfApp5
 
         private void buttonEkle_Click(object sender, RoutedEventArgs e)
         {
-            var item = context.Definations.Where(x => x.DefValue == comboMeasurementUnit.Text).FirstOrDefault();
-            if (item==null)
+            var item = context.Definations.Where(x => x.DefValue == comboMeasurementUnit.Text && x.DefType==(int)Definition.Unit).SingleOrDefault();
+            if (item == null)
             {
-                if (!string.IsNullOrEmpty(comboMeasurementUnit.Text))
+                if (string.IsNullOrEmpty(comboMeasurementUnit.Text))
                 {
-                    var newItem = new Defination();
-                    newItem.DefType = 2;
-                    newItem.DefValue = comboMeasurementUnit.Text;
-                    context.Definations.Add(newItem);
-                    context.SaveChanges();
-                    MessageBox.Show(comboMeasurementUnit.Text + "Birimi Başarıyla Eklendi");
-                    comboShow();
+                    MessageBox.Show("Ölçü Birimi Boş Bırakılamaz!");
                 }
                 else
                 {
-                    MessageBox.Show( "Boş Kayıt Eklenemez!");
+                    Defination newItem = new Defination();
+                    newItem.DefType = (int)Definition.Unit;
+                    newItem.DefValue = comboMeasurementUnit.Text;
+                    context.Definations.Add(newItem);
+                    context.SaveChanges();
+                    MessageBox.Show("Ekleme İşlemi Başarılı");
+                    comboShow();
                 }
             }
             else
             {
                 MessageBox.Show(comboMeasurementUnit.Text + " Kayıdı Zaten Var!");
             }
-
-            //context.Definations.Add(item);
         }
 
         private void buttonSil_Click(object sender, RoutedEventArgs e)
         {
-            if (!string.IsNullOrEmpty( comboMeasurementUnit.Text))
+            if (!string.IsNullOrEmpty(comboMeasurementUnit.Text))
             {
-                var item = context.Definations.Where(x => x.Id == Convert.ToInt32(comboMeasurementUnit.SelectedValue)).FirstOrDefault();
-                context.Definations.Remove(item);
-                context.SaveChanges();
-                string name = comboMeasurementUnit.Text;
-                comboShow();
-                MessageBox.Show(name +" Ölçü Birimini Silme İşlemi Başarılı.");
+                var item = context.Definations.Where(x => x.DefValue == comboMeasurementUnit.Text && x.DefType == (int)Definition.Unit).FirstOrDefault();
+                if (item != null)
+                {
+                    context.Definations.Remove(item);
+                    context.SaveChanges();
+                    string name = comboMeasurementUnit.Text;
+                    comboShow();
+                    MessageBox.Show(name + " Ölçü Birimini Silme İşlemi Başarılı.");
+                }
+                else
+                {
+                    MessageBox.Show(comboMeasurementUnit.Text + " Adında Bir Ölçü Birimi Bulunmamakta!");
+                }
             }
         }
 
         private void buttonGuncelle_Click(object sender, RoutedEventArgs e)
         {
-            if (comboMeasurementUnit.SelectedValue!=null)
+            var item = context.Definations.Where(x => x.DefValue == comboMeasurementUnit.Text && x.DefType == (int)Definition.Unit).FirstOrDefault();
+            if (!string.IsNullOrEmpty(comboMeasurementUnit.Text))
             {
-                MeasurementUnitUpdate mu = new MeasurementUnitUpdate();
-                mu.comboMeasurementName1.ItemsSource = comboMeasurementUnit.ItemsSource;
-                mu.comboMeasurementName1.DisplayMemberPath = comboMeasurementUnit.DisplayMemberPath;
-                mu.comboMeasurementName1.SelectedValuePath = comboMeasurementUnit.SelectedValuePath;
-                mu.comboMeasurementName1.SelectedValue = comboMeasurementUnit.SelectedValue;
-                this.Close();
-                mu.ShowDialog();
-            }
-            else
-            {
-                MessageBox.Show("Güncelleme Yapmadan Önce Güncellemek İstediğiniz Dağıtıcıyı Seçmelisiniz.");
+                if (item != null)
+                {
+                    MeasurementUnitUpdate measurement = new MeasurementUnitUpdate();
+                    measurement.lblName.Content = item.DefValue;
+                    this.Close();
+                    measurement.ShowDialog();
+                }
+                else
+                {
+                    MessageBox.Show(comboMeasurementUnit.Text + " Adında Bir Ölçü Birimi Bulunmamakta!");
+                }
             }
         }
     }
